# docs/swagger.yaml
definitions:
...
  dto.AddHeartsRequest:
    properties:
      amount:
        maximum: 10
        minimum: 1
        type: integer
      source:
        description: '"ad", "purchase", "daily_reset"'
        enum:
        - ad
        - purchase
        - daily_reset
        type: string
    required:
    - amount
    - source
    type: object
...
  dto.AuditLogResponse:
    properties:
      limit:
        example: 20
        type: integer
      logs:
        items:
          $ref: '#/definitions/dto.AuthAuditLog'
        type: array
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
    type: object
...
  dto.CollectionResponse:
    properties:
      achievements:
        items:
          $ref: '#/definitions/dto.AchievementResponse'
        type: array
      characters:
        $ref: '#/definitions/dto.CharacterCollectionResponse'
      stats:
        $ref: '#/definitions/dto.CollectionStatsResponse'
    type: object
...
  dto.CompleteLessonRequest:
    properties:
      lesson_id:
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
      time_spent:
        minimum: 1
        type: integer
    required:
    - lesson_id
    - score
    - time_spent
    type: object
...
  dto.DeviceInfo:
    properties:
      browser:
        example: Safari
        type: string
      id:
        example: dev_123456789
        type: string
      ip:
        example: 192.168.1.1
        type: string
      is_trusted:
        example: true
        type: boolean
      last_used:
        example: "2023-01-15T10:30:00Z"
        type: string
      name:
        example: iPhone 14
        type: string
      os:
        example: iOS 16.0
        type: string
      type:
        example: mobile
        type: string
    type: object
...
  dto.HeartStatusResponse:
    properties:
      ads_watched_today:
        type: integer
      can_watch_ad:
        type: boolean
      hearts:
        type: integer
      max_hearts:
        type: integer
      next_reset_time:
        type: string
    type: object
...
  dto.LessonAccessResponse:
    properties:
      can_access:
        type: boolean
      hearts_needed:
        type: integer
      reason:
        type: string
    type: object
...
  dto.TrustDeviceRequest:
    properties:
      device_id:
        example: device_12345
        type: string
      trust:
        example: true
        type: boolean
    required:
    - device_id
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      email:
        example: newemail@example.com
        type: string
      username:
        example: newusername
        maxLength: 30
        minLength: 3
        type: string
    type: object
  dto.UpdateSecuritySettingsRequest:
    properties:
      login_notifications:
        example: true
        type: boolean
      session_timeout:
        example: 720
        maximum: 43200
        minimum: 300
        type: integer
    type: object
...
  dto.UserProfileResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: usr_123456789
        type: string
      is_active:
        example: true
        type: boolean
      last_login_at:
        example: "2023-01-15T10:30:00Z"
        type: string
      last_login_ip:
        example: 192.168.1.1
        type: string
      role:
        example: user
        type: string
      stats:
        $ref: '#/definitions/dto.UserStats'
      username:
        example: johndoe
        type: string
    type: object
  dto.UserProgressResponse:
    properties:
      completed_lessons:
        items:
          type: string
        type: array
      hearts:
        type: integer
      last_activity:
        type: string
      last_heart_reset:
        type: string
      level:
        type: integer
      max_hearts:
        type: integer
      recent_achievements:
        items:
          $ref: '#/definitions/dto.AchievementResponse'
        type: array
      spirit:
        $ref: '#/definitions/dto.SpiritResponse'
      streak:
        type: integer
      total_play_time:
        type: integer
      unlocked_characters:
        items:
          type: string
        type: array
      user_id:
        type: string
      xp:
        type: integer
      xp_to_next_level:
        type: integer
    type: object
...
  shared.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/user/audit-logs:
    get:
      consumes:
      - application/json
      description: Get user authentication audit logs
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuditLogResponse'
              type: object
      security:
      - Bearer: []
      summary: Get audit logs
      tags:
      - user
  /api/v1/user/collection:
    get:
      consumes:
      - application/json
      description: Get user collection
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CollectionResponse'
              type: object
      security:
      - Bearer: []
      summary: Get user collection
      tags:
      - user
  /api/v1/user/devices:
    get:
      consumes:
      - application/json
      description: Get list of all user's trusted devices
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.DeviceInfo'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get user devices
      tags:
      - user
  /api/v1/user/devices/{deviceId}:
    delete:
      consumes:
      - application/json
      description: Remove a device from user's device list
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Device ID
        in: path
        name: deviceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.Response'
      security:
      - Bearer: []
      summary: Remove user device
      tags:
      - user
  /api/v1/user/devices/{deviceId}/trust:
    put:
      consumes:
      - application/json
      description: Trust or untrust a specific device
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Device ID
        in: path
        name: deviceId
        required: true
        type: string
      - description: Trust status
        in: body
        name: trustRequest
        required: true
        schema:
          $ref: '#/definitions/dto.TrustDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.Response'
      security:
      - Bearer: []
      summary: Update device trust status
      tags:
      - user
  /api/v1/user/hearts:
    get:
      consumes:
      - application/json
      description: Get user heart status
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HeartStatusResponse'
              type: object
      security:
      - Bearer: []
      summary: Get user heart status
      tags:
      - user
  /api/v1/user/hearts/add:
    post:
      consumes:
      - application/json
      description: Add user hearts
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add hearts request
        in: body
        name: addRequest
        required: true
        schema:
          $ref: '#/definitions/dto.AddHeartsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HeartStatusResponse'
              type: object
      security:
      - Bearer: []
      summary: Add user hearts
      tags:
      - user
  /api/v1/user/hearts/lose:
    post:
      consumes:
      - application/json
      description: Lose user heart
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HeartStatusResponse'
              type: object
      security:
      - Bearer: []
      summary: Lose user heart
      tags:
      - user
  /api/v1/user/initialize:
    post:
      consumes:
      - application/json
      description: Initialize user profile
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User profile
        in: body
        name: initializeRequest
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserProgressResponse'
              type: object
      security:
      - Bearer: []
      summary: Initialize user profile
      tags:
      - user
  /api/v1/user/lesson/{lessonId}/access:
    get:
      consumes:
      - application/json
      description: Check user lesson access
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Lesson ID
        in: path
        name: lessonId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LessonAccessResponse'
              type: object
      security:
      - Bearer: []
      summary: Check user lesson access
      tags:
      - user
  /api/v1/user/lesson/complete:
    post:
      consumes:
      - application/json
      description: Complete user lesson
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Complete lesson request
        in: body
        name: completeRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CompleteLessonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserProgressResponse'
              type: object
      security:
      - Bearer: []
      summary: Complete user lesson
      tags:
      - user
  /api/v1/user/profile:
    get:
      consumes:
      - application/json
      description: Get user profile
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserProfileResponse'
              type: object
      security:
      - Bearer: []
      summary: Get user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user profile
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User profile
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserProfileResponse'
              type: object
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - user
  /api/v1/user/progress:
    get:
      consumes:
      - application/json
      description: Get user progress
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserProgressResponse'
              type: object
      security:
      - Bearer: []
      summary: Get user progress
      tags:
      - user
  /api/v1/user/security:
    get:
      consumes:
      - application/json
      description: Get user security settings
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SecuritySettings'
              type: object
      security:
      - Bearer: []
      summary: Get security settings
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user security settings
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Security settings
        in: body
        name: updateRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSecuritySettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SecuritySettings'
              type: object
      security:
      - Bearer: []
      summary: Update security settings
      tags:
      - user
  /api/v1/user/sessions:
    get:
      consumes:
      - application/json
      description: Get list of active user sessions
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.SessionListResponse'
              type: object
      security:
      - Bearer: []
      summary: Get user sessions
      tags:
      - user
  /api/v1/user/sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Revoke a specific user session
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Revoke user session
      tags:
      - user
  /api/v1/user/share:
    post:
      consumes:
      - application/json
      description: Share achievement
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Share request
        in: body
        name: shareRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ShareResponse'
              type: object
      security:
      - Bearer: []
      summary: Share achievement
      tags:
      - user
...
swagger: "2.0"